refactor(conventions): split detectors into per-family files and decompose
Break conventions.go (1553 lines) into orchestration plus cohesive
per-family files: detectors_generic.go, detectors_go.go,
detectors_framework.go (Rails + React idioms), detectors_arch.go, and
helpers.go. conventions.go keeps Detect and the loaders.

Decompose all six gocyclo/gocognit-over-threshold functions into named
helpers (not abstractions), preserving Detect's exact dispatch order:
detectTesting, Detect, detectGoMiddleware, detectScopes,
detectArchitectureLayers, detectDesignPatterns. Every //nolint entry in
the package is retired.

Cover-before-decompose: detectGoInterfaces 45.8% to 100% with a real
interface fixture (no latent bug found, the gap was a missing fixture);
detectTesting and detectGoMiddleware to 100%; loader query-error paths
covered via a closed-DB harness. Package coverage 91.0% to 95.4%.

Behavior preserved: verified byte-identical (under a normalized oracle
that strips pre-existing tie-order nondeterminism) against the maket
index across the unfiltered, models, and controllers queries, through
both the pure move and all decompositions.
